Fix item template targetoutput path
An item template is supposed to live inside a project. The way `TargetOutput` was resolved previously kept the basename of the folder specified resulting in an incorrect path for the item template. For example, if a user specified `c:/users/james/foo` the item template would create `c:/users/james/foo/foo/item` when it should create `c:/users/james/foo/item`. This has been fixed by removing the basename if the template is an item.